agent/agent/main.py
from google.adk.agents import LlmAgent
from google.adk.tools import google_search
from pydantic import BaseModel
from typing import Dict, Any, List
import json
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class ImageAnalysisResult(BaseModel):
    """Result from image analysis tools"""
    objects_detected: List[str]
    scene_description: str
    road_detected: bool
    obstacles: List[str]
    confidence_score: float

class NavigationData(BaseModel):
    """Navigation context data"""
    current_location: Dict[str, float]  # lat, lng
    destination: Dict[str, float] = None
    route_instructions: List[str] = []
    nearby_landmarks: List[str] = []

def analyze_scene_with_gemini(image_data: bytes) -> str:
    """
    Analyze scene using Gemini multimodal vision.
    This would integrate with your Gemini Vision processing pipeline.
    """
    # Placeholder - integrate with your Gemini Vision API
    return """Scene: Urban street with pedestrians and vehicles. 
    Clear sidewalk ahead. Traffic light showing green. 
    No immediate obstacles detected."""

def detect_objects_yolo(image_data: bytes) -> List[str]:
    """
    Detect objects using YOLO model.
    This would integrate with your YOLO Object Detection pipeline.
    """
    # Placeholder - integrate with your YOLO detection
    return ["person", "car", "traffic_light", "sidewalk", "building"]

def segment_road_path(image_data: bytes) -> Dict[str, Any]:
    """
    Detect road/path using SegNet.
    This would integrate with your SegNet Image Segmentation pipeline.
    """
    # Placeholder - integrate with your SegNet model
    return {
        "road_detected": True,
        "clear_path": True,
        "path_direction": "straight",
        "obstacles_in_path": []
    }

def get_navigation_context(gps_data: Dict[str, float]) -> str:
    """
    Get navigation context from GPS data.
    This would integrate with your GPS Handler and Navigation Tools.
    """
    # Placeholder - integrate with your GPS processing
    return f"Current location: {gps_data.get('lat', 0)}, {gps_data.get('lng', 0)}. Clear path ahead."

def generate_audio_response(message: str) -> str:
    """
    Generate audio response for the user.
    This would integrate with your Audio Response Tools.
    """
    # Placeholder - integrate with your audio generation
    return f"Audio response generated: {message}"

# Define the ReAct AI Agent
react_agent = LlmAgent(
    model="gemini-2.5-flash",
    name="smart_glasses_react_agent",
    description="AI decision engine for smart glasses navigation and scene understanding",
    instruction="""You are the ReAct AI Agent for a smart glasses navigation system. Your role is to:

1. ANALYZE incoming visual, audio, and GPS data
2. REASON about the user's context and needs
3. ACT by providing appropriate navigation instructions

Available tools and context:
- Scene analysis from Gemini Vision (detailed scene understanding)
- Object detection from YOLO (specific object identification)
- Path segmentation from SegNet (road/path detection)
- GPS navigation data (location and routing)
- Audio response generation (for user feedback)

Your responses should be:
- CONCISE and ACTIONABLE for audio delivery
- SAFETY-FOCUSED (warn about obstacles, traffic)
- CONTEXT-AWARE (consider user's location and movement)
- HELPFUL for navigation decisions

When processing input:
1. First analyze what information you have (visual scene, GPS, user intent)
2. Reason about potential actions or warnings needed
3. Provide clear, brief navigation instructions

Example flow:
- Input: "User walking, scene shows intersection ahead"
- Analysis: "Intersection detected, traffic light visible, pedestrian crossing available"
- Action: "Intersection ahead. Traffic light is green. Safe to cross using crosswalk."

Keep responses under 15 seconds of spoken content for optimal user experience.""",
    tools=[
        analyze_scene_with_gemini,
        detect_objects_yolo, 
        segment_road_path,
        get_navigation_context,
        generate_audio_response,
        google_search  # For additional context when needed
    ]
)

# Root agent (main entry point)
root_agent = react_agent 